+++ Forge/src/main/java/net/merchantpug/bovinesandbuttercups/client/resources/ModFilePackResources.java
@@ -31,16 +31,18 @@ import java.nio.file.Path;
 
 public class ModFilePackResources extends PathPackResources {
     protected final IModFile modFile;
+    protected final String sourcePath;
 
-        public ModFilePackResources(String name, IModFile modFile, Path source) {
-        super(name, true, source);
+    public ModFilePackResources(String name, IModFile modFile, String sourcePath) {
+        super(name, true, modFile.findResource(sourcePath));
         this.modFile = modFile;
+        this.sourcePath = sourcePath;
     }
 
     @Override
     protected Path resolve(String... paths) {
         String[] allPaths = new String[paths.length + 1];
-        allPaths[0] = getSource().toString();
+        allPaths[0] = sourcePath;
         System.arraycopy(paths, 0, allPaths, 1, paths.length);
         return modFile.findResource(allPaths);
     }