+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Bugfixes
 - Fixed Mojang Textures and No Grass Back built-in resource packs having an outdated path for their textures.
+- [FORGE] Fixed broken built-in resource pack implementation.
 - [FORGE] Fixed reload command failing due to faulty packet code.
 - [FORGE] Fixed Forge thinking that Bovines always has an update by changing the version schema.